Split DatasetJudgementServiceTest.kt to comply with Sonar length restriction
